+++ client/src/book-editor/BookEditor.js
@@ -1,9 +1,18 @@
 import React, { useState, useEffect, useRef } from 'react'
 import { useParams, Link } from 'react-router-dom'
-import { Navbar, Dropdown } from 'react-bootstrap'
+import { Navbar, Dropdown, Card, Button } from 'react-bootstrap'
 import { SimpleDropdown } from '../components/simple-dropdown/SimpleDropdown'
 import io from 'socket.io-client'
 import { LoadingHome } from '../app/App'
+import AceEditor from 'react-ace'
+import './BookEditor.css'
+
+// Ace editor imports
+import "ace-builds/src-noconflict/mode-scala"
+import "ace-builds/src-noconflict/theme-textmate"
+import "ace-builds/src-min-noconflict/ext-searchbox";
+import "ace-builds/src-min-noconflict/ext-language_tools";
+import "ace-builds/src-noconflict/snippets/scala";
 
 function EditorNavbar(props) {
     const copyAll = () => {
@@ -40,6 +49,91 @@ function EditorNavbar(props) {
     )
 }
 
+function CommandChunk(props) {
+    const statusList = {
+        waiting: { labal: 'Waiting connection...' },
+        untouched: { label: '' },
+        running: { label: 'Running...' },
+        done: { label: 'Done' }
+    }
+    const [status, changeStatus] = useState(null)
+
+    const codeChange = value => {
+        // TODO something
+        console.log(value)
+    }
+
+    const run = () => {
+        // TODO something
+        console.log('executou esse...')
+    }
+
+    const runAllAbove = () => {
+        // TODO something
+        console.log('executou todos acima...')
+    }
+
+    return (
+        <Card className="m-3 shadow command-block">
+            <Card.Header className="d-flex">
+                <div>
+                    [Chunk {props.index}] {props.command.name}
+                </div>
+                <span className="flex-grow-1"></span>
+                <span>
+                    <i className="fa fa-ellipsis-v"></i>
+                </span>
+            </Card.Header>
+            <Card.Body as={AceEditor}
+                name={`Chunk-${props.index}`}
+                value={props.command.command}
+                onChange={codeChange}
+                mode="scala"
+                theme="textmate"
+                setOptions={{
+                    maxLines: Infinity,
+                    minLines: 5,
+                    showPrintMargin: false,
+                    enableBasicAutocompletion: true,
+                    enableLiveAutocompletion: false,
+                    enableSnippets: true,
+                    showLineNumbers: false,
+                    useSoftTabs: true,
+                    tabSize: 2
+                }}
+                commands={[
+                    {
+                        name: 'run',
+                        bindKey: {
+                            win: 'Ctrl-Enter',
+                            mac: 'Command-Enter',
+                        },
+                        exec: run
+                    },
+                    {
+                        name: 'run-all',
+                        bindKey: {
+                            win: 'Control-Shift-Enter',
+                            mac: 'Command-shift-enter'
+                        },
+                        exec: runAllAbove
+                    }
+                ]}
+                style={{
+                    width: '100%'
+                }}
+            ></Card.Body>
+            <Card.Footer className="d-flex">
+                <span>{status ? status.label : null}</span>
+                <span className="flex-grow-1"></span>
+                <Button disabled onClick={run}>
+                    <i className="fa fa-play"></i>
+                </Button>
+            </Card.Footer>
+        </Card>
+    )
+}
+
 export function BookEditor(props) {
     const { bookId } = useParams()
     const [loading, setLoading] = useState(true)
@@ -66,6 +160,7 @@ export function BookEditor(props) {
     return (
         <LoadingHome loading={loading}>
             <EditorNavbar book={book}/>
+            { book ? book.commands.map((command, index) => <CommandChunk command={command} key={index} index={index} bookSocket={bookSocketRef.current} />) : null }
         </LoadingHome>
     )
 }